+++ src/AnimalPanel.java
